#2 Refactor: API 모듈을 분리하여 배럴 패턴 적용
