+++ src/services/auth/index.ts
@@ -1,0 +1,2 @@
+export { default as login } from './login';
+export { default as logout } from './logout';
