--- src/services/signup.ts
@@ -1,16 +1,0 @@
-import axiosInstance from '@/api/axiosInstance';
-import { SignupData } from '@/types/d';
-
-const signup = async (data: Omit<SignupData, 'confirmPassword'>) => {
-  try {
-    const response = await axiosInstance.post('/users/signup', data);
-    return response.data.user_id; // 성공 시 반환되는 user_id 반환
-  } catch (error: any) {
-    if (error.response) {
-      throw error.response; // API 호출 실패 시, 오류 응답을 그대로 던집니다.
-    }
-    throw error;
-  }
-};
-
-export default signup;
