+++ src/services/users/index.ts
@@ -1,0 +1,1 @@
+export { default as signup } from './signup';
